requirements.txt
aiogram==2.22.1
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.6.1
APScheduler==3.6.3
apturl==0.5.2
async-timeout==4.0.2
attrs==22.1.0
autocommand==2.2.1
Babel==2.9.1
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
cachetools==4.2.2
certifi==2022.6.15
chardet==4.0.0
charset-normalizer==2.1.1
cheroot==8.6.0
CherryPy==18.8.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
decorator==5.1.1
defer==1.0.6
distro==1.7.0
distro-info===1.1build1
dnspython==2.2.1
duplicity==0.8.21
email-validator==1.2.1
environs==9.5.0
fasteners==0.14.1
frozenlist==1.3.1
future==0.18.2
google-api-python-client==1.7.11
google-auth==1.5.1
google-auth-httplib2==0.1.0
greenlet==1.1.2
h11==0.12.0
h2==4.1.0
hpack==4.0.0
httpcore==0.15.0
httplib2==0.20.2
httpx==0.23.0
hypercorn==0.14.1
hyperframe==6.0.1
idna==3.3
importlib-metadata==4.6.4
inflect==6.0.0
iniconfig==1.1.1
ipaddress==1.0.23
jaraco.classes==3.2.2
jaraco.collections==3.5.2
jaraco.context==4.1.2
jaraco.functools==3.5.1
jaraco.text==3.9.1
jeepney==0.7.1
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
loguru==0.6.0
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
marshmallow==3.17.1
monotonic==1.6
more-itertools==8.14.0
multidict==6.0.2
netifaces==0.11.0
oauth2client==4.1.3
oauthlib==3.2.0
olefile==0.46
packaging==21.3
paramiko==2.9.3
pexpect==4.8.0
phonenumbers==8.12.54
Pillow==9.0.1
pluggy==1.0.0
portend==3.1.0
priority==2.0.0
protobuf==3.12.4
psycopg2-binary==2.9.3
ptyprocess==0.7.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycairo==1.20.1
pycups==2.0.1
pydantic==1.9.2
PyDrive==1.3.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==3.0.9
pyQiwiP2P==2.0.6
pyRFC3339==1.1
pytest==7.1.2
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-dotenv==0.20.0
python-telegram-bot==13.14
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.28.1
response-report==1.1.1
rfc3986==1.5.0
rsa==4.8
SecretStorage==3.3.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.31
starlette==0.20.4
systemd-python==234
tempora==5.0.2
toml==0.10.2
tomli==2.0.1
tornado==6.1
typing==3.7.4.3
typing_extensions==4.3.0
tzdata==2022.5
tzlocal==4.2
ubuntu-advantage-tools==27.11.3
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.26.12
usb-creator==0.3.7
validators==0.20.0
wadllib==1.3.6
wsproto==1.2.0
xdg==5
xkit==0.0.0
yarl==1.8.1
zc.lockfile==2.0
zipp==1.0.0
